+++ modules/extensions-spring-support/src/test/resources/typed-value-config.xml
@@ -88,4 +88,11 @@
         <typed-value:source stringValue="#[payload]"/>
         <set-payload value="string" mimeType="application/json" encoding="UTF-8"/>
     </flow>
+
+    <flow name="typedValueForStringInsidePojo">
+        <set-payload value="string" mimeType="application/json" encoding="UTF-8"/>
+        <typed-value:typed-value-for-string-inside-pojo>
+            <typed-value:differed-door address="#[payload]"/>
+        </typed-value:typed-value-for-string-inside-pojo>
+    </flow>
 </mule>
